libmpathutil: move devt2devname() to libmultipath
devt2devname is under the default license and can be moved to
libmultipath.
